Move GLee header to system-like directory
